Huffman: fix bug with bitshift more than or equal the width
Shifting an int 32 bits is udnefined

On gpu it seems that shifting an int with 32 bits sets it to zero - on
cpu it remained unchanged (on my machines at least)

ISO/IEC 9899:1999 6.5.7 Bitwise shift operators ¶3

The integer promotions are performed on each of the operands. The
type of the result is that of the promoted left operand. If the value of
the right operand is negative or is greater than or equal to the width
of the promoted left operand, the behavior is undefined.